+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# IEATools 0.1.46 (2012-02-01)
+
 * `calc_tidy_iea_df_balances` function updated
   so that it also accounts for balancing flows 
   when checking balances.